backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
sentence-transformers
rapidfuzz
python-multipart
